add doubleclick to rotate plot on grid
this is done through a flag and timeout in game object.
the flag is either false or set to a plot id.
the timeout resets the flag to false after 400ms.
when a plot is clicked, it checks if it's currently flagged for a double
click.
if it is flagged then it rotates, if it isn't flagged, it flags itself.

flagging a plot clears the previous flag and timeout, ensuring that
clicking plots rapidly doesn't create a stack of timeouts.

finally, if the plot was already selected, and the last click was too slow
for a double click, then try to move the plot to the new click location.